Magellan: removed old display code, moved around pacakges
