+++ plugins/Magellan/src/main/java/org/micromanager/magellan/gui/GUI.java
@@ -62,7 +62,6 @@ import org.micromanager.magellan.channels.ColorEditor;
 import org.micromanager.magellan.channels.ColorRenderer;
 import org.micromanager.magellan.coordinates.MagellanAffineUtils;
 import org.micromanager.magellan.main.Magellan;
-import org.micromanager.magellan.misc.ExactlyOneRowSelectionModel;
 import org.micromanager.magellan.misc.GlobalSettings;
 import org.micromanager.magellan.misc.JavaUtils;
 import org.micromanager.magellan.misc.LoadedAcquisitionData;
