+++ plugins/Magellan/src/main/java/org/micromanager/magellan/imagedisplaynew/ColorableScrollbarUI.java
@@ -15,7 +15,7 @@
 //               INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES.
 //
 
-package org.micromanager.magellan.imagedisplay;
+package org.micromanager.magellan.imagedisplaynew;
 
 import java.awt.Color;
 import java.awt.Graphics;
