+++ plugins/Magellan/src/main/java/org/micromanager/magellan/imagedisplaynew/MagellanDataViewCoords.java
@@ -7,7 +7,6 @@ package org.micromanager.magellan.imagedisplaynew;
 
 import java.awt.geom.Point2D;
 import java.util.HashMap;
-import sun.awt.image.ImageCache;
 
 /**
  *
