+++ plugins/Magellan/src/main/java/org/micromanager/magellan/imagedisplaynew/MetadataPanel.form
