+++ plugins/Magellan/src/main/java/org/micromanager/magellan/imagedisplaynew/MetadataPanel.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.micromanager.magellan.imagedisplay;
+package org.micromanager.magellan.imagedisplaynew;
 
 import ij.ImagePlus;
 import java.util.Timer;
